move executables-related commands to executables from commands cache
since they're used in paths cache as well and paths cache can't import commands_cache due to circular imports
